fix names and function
