+++ js/random-integer.js
@@ -1,4 +1,10 @@
-import { getRandomInteger } from './get-random-integer';
+
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
 
 function createRandomIdFromRangeGenerator(min, max) {
   const previousValues = [];
@@ -16,4 +22,4 @@ function createRandomIdFromRangeGenerator(min, max) {
   };
 }
 
-export { createRandomIdFromRangeGenerator };
+export { createRandomIdFromRangeGenerator, getRandomInteger };
